Move src/api -> src/backend, scripts -> backend/scripts
