+++ package.json
@@ -7,8 +7,8 @@
     "type-check": "tsc --noEmit",
     "test": "ava",
     "test:watch": "ava --watch",
-    "db:migrate": "npx tsx src/scripts/migrate-database.ts",
-    "db:reset": "npx tsx src/scripts/reset-database.ts",
+    "db:migrate": "npx tsx src/backend/scripts/migrate-database.ts",
+    "db:reset": "npx tsx src/backend/scripts/reset-database.ts",
     "prepare": "husky install",
     "format": "prettier --write --ignore-unknown .",
     "format:check": "prettier --check --ignore-unknown .",
