+++ src/backend/context.ts
@@ -1,5 +1,5 @@
 import { CreateNextContextOptions } from "@trpc/server/adapters/next";
-import { getDb } from "api/db/lib/get-db";
+import { getDb } from "backend/db/lib/get-db";
 import { Env } from "./lib/types";
 
 export const createContext = (options: CreateNextContextOptions) => {
