+++ src/backend/db/lib/get-db.ts
@@ -1,5 +1,5 @@
 import { Kysely, PostgresDialect } from "kysely";
-import { KyselyDatabaseInstance, KyselySchema } from "api/lib";
+import { KyselyDatabaseInstance, KyselySchema } from "backend/lib";
 import { Pool } from "pg";
 
 type DatabaseInstanceAndPool = { db: KyselyDatabaseInstance; pool: Pool };
