+++ src/backend/db/lib/migrate.ts
@@ -2,7 +2,7 @@ import * as postgresMigrations from "postgres-migrations";
 import * as zg from "zapatos/generate";
 import { Pool } from "pg";
 import path from "node:path";
-import { logger } from "api/lib";
+import { logger } from "backend/lib";
 
 interface MigrateOptions {
   silent?: boolean;
@@ -13,8 +13,11 @@ export const migrate = async (
   databaseUrl: string,
   { silent = false, shouldGenerateTypes = true }: MigrateOptions = {}
 ) => {
-  const migrationsDirectory = path.join(process.cwd(), "src/api/db/migrations");
-  const zapatosDirectory = path.join(process.cwd(), "src/api/db");
+  const migrationsDirectory = path.join(
+    process.cwd(),
+    "src/backend/db/migrations"
+  );
+  const zapatosDirectory = path.join(process.cwd(), "src/backend/db");
 
   if (!silent) {
     logger.info("Migrating database...");
