+++ src/backend/db/lib/reset.ts
@@ -1,6 +1,6 @@
 import { Client } from "pg";
 import { parse } from "pg-connection-string";
-import { logger } from "api/lib";
+import { logger } from "backend/lib";
 import { migrate } from "./migrate";
 
 export const reset = async (databaseUrl: string) => {
