+++ src/client/lib/trpc.ts
@@ -1,4 +1,4 @@
 import { createTRPCReact } from "@trpc/react-query";
-import { AppRouter } from "api/routers/_app";
+import { AppRouter } from "backend/routers/_app";
 
 export const trpc = createTRPCReact<AppRouter>();
