+++ src/pages/api/trpc/[trpc].ts
@@ -1,6 +1,6 @@
 import * as trpcNext from "@trpc/server/adapters/next";
-import { createContext } from "api/context";
-import { trpcRouter } from "api/routers/_app";
+import { createContext } from "backend/context";
+import { trpcRouter } from "backend/routers/_app";
 
 export default trpcNext.createNextApiHandler({
   router: trpcRouter,
